Add traits to make defining facades easier in most cases.
The new SimpleFacade and MutableFacade correspond to the two most
obvious use cases (building from immutable and mutable collections).
The only real need to define a Facade directly is when it is more
efficient to work with other types than convert from these.

At this point only Argonaut's facade is manual. This commit also adds
a Rojoma facade.