Fix unintuitive handling of Unicode surrogates.
This fixes issue #12.

Jawn had some special code to parse UTF-16 characters which are
"high surrogate" (in the range DBFF-D800). The rationale was the
next character might be interpreted incorrectly. However, since
"low surrogates" are required to be in the range DFFF-DC00, there
is no real potential for the "standard" parsing to be wrong here.

This change should make parsing strings a tiny bit faster, and
will also help us parse strings that aren't really valid Unicode
(such as in the example).