support Play 2.4
Play 2.4 changed the constructor of `JsObject` to take a `Map` instead
of `Seq`. The latter variant is now an `apply` on the companion
object, losing binary compatibility.  Play 2.3 will require an older
version of `jawn-play` without this change.

The `Parser` can reuse the SimpleFacade defaults for all Json value
types.