Make ChannelParser able to deal with very big strings.
Previously strings that were outside of the buffer size (1M)
would not be handled correctly. This ended up causing a real
issue for someone.

The new strategy is to grow the buffer size whenever long strings
are encountered. This doesn't seem to slow things down too much
and it seems better to handle those situations.

Strings that are >1G will still fail under this strategy, since
we won't be able to allocate a large-enough array -- but there
will always be an upper limit, and this commit moves that to
a reasonable place (IMO).